FIx merge insert sort

+++ module09/ex02/srcs/main.cpp
@@ -39,7 +39,7 @@ int main(int argc, char *argv[])
     clock_t beginVector = clock();
     std::vector<int> sorted_vector = pmergeMe.sortVector();
     clock_t endVector = clock();
-    double elapsedSecsVector = double(endVector - beginVector) / CLOCKS_PER_SEC;
+    double elapsedSecsVector = double(endVector - beginVector) / double(CLOCKS_PER_SEC);
     std::cout << "After: ";
     printVector(sorted_vector);
     std::cout << "Time to process a range of " << pmergeMe.getVector().size() << " elements with std::vector : " << elapsedSecsVector << " seconds" << std::endl << std::endl << std::endl;
@@ -49,7 +49,7 @@ int main(int argc, char *argv[])
     clock_t beginList = clock();
     std::list<int> sorted_list = pmergeMe.sortList();
     clock_t endList = clock();
-    double elapsedSecsList = double(endList - beginList) / CLOCKS_PER_SEC;
+    double elapsedSecsList = double(endList - beginList) / double(CLOCKS_PER_SEC);
     std::cout << "After: ";
     printList(sorted_list);
     std::cout << "Time to process a range of " << pmergeMe.getList().size() << " elements with std::list : " << elapsedSecsList << " seconds" << std::endl;
